feat(T035): implement workspace tabs state and models
Agent-Logs-Url: https://github.com/nelrsf/UI-frame/sessions/e228ef02-b989-4221-8071-7d1f6c546cc2
