+++ src/app/core/state/workspace/index.ts
@@ -1,0 +1,3 @@
+export * from './workspace.actions';
+export * from './workspace.reducer';
+export * from './workspace.selectors';
